crm_audio_pipeline/requirements.txt
openai-whisper==20240930
pymongo
grog
pydub
regex
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
requests
cloudinary
